Fixing missing torch package in TF images
